feat: FastApi 서버와 통신 오류 시 예외 처리 #22
